Fix a pointless cast from uintptr_t to long.

+++ usr.sbin/tcpdrop/tcpdrop.c
@@ -235,7 +235,7 @@ tcpdropall(const char *stack, int state)
 	head = getxpcblist("net.inet.tcp.pcblist");
 
 #define	XINP_NEXT(xinp)							\
-	((struct xinpgen *)(long)((uintptr_t)(xinp) + (xinp)->xig_len))
+	((struct xinpgen *)((uintptr_t)(xinp) + (xinp)->xig_len))
 
 	for (xinp = XINP_NEXT(head); xinp->xig_len > sizeof *xinp;
 	    xinp = XINP_NEXT(xinp)) {
@@ -384,3 +384,11 @@ usage(void)
 "       tcpdrop [-l] -S stack -s state\n");
 	exit(1);
 }
+// CHERI CHANGES START
+// {
+//   "updated": 20180907,
+//   "changes": [
+//     "pointer_integrity"
+//   ]
+// }
+// CHERI CHANGES END
